Refactored the get requests. started css page on the front end
