+++ client/src/Pages/Threads/ThreadsPage.js
@@ -1,0 +1,25 @@
+// PAGE ELEMENTS
+import { React } from "react";
+import Navigation from "../../components/Navigation/Navigation.js";
+
+// STYLES
+
+// API
+import api from "../../Api";
+
+function ThreadPage() {
+  return (
+    <>
+      <Navigation />
+      <div className={styles.body_inner}></div>
+    </>
+  );
+}
+export default ThreadPage;
+
+/*
+-------------------------------------
+CREDITS
+-------------------------------------
+- Developers institute
+ */
