+++ server/forum/postThread.repository.js
@@ -9,8 +9,6 @@ const pool = require("../db");
 const createThreadSQL = `INSERT INTO thread (thread_title, thread_text, updated_at, topic_id, author, created_at) VALUES($1, $2, $3, $4, $5, $6) 
 RETURNING thread_id, thread_title, thread_text, created_at, updated_at, topic_id, author`;
 
-const getPostSQL = `SELECT * FROM thread`;
-
 /* 
 -----------------------------------
        REPOSITORY FUNCTIONS
@@ -43,15 +41,4 @@ const createThread = async (
     throw Error(error);
   }
 };
-
-const getAllForumThreads = async () => {
-  try {
-    const Pool = await pool();
-
-    const result = await Pool.query(getPostSQL);
-    return result.rows;
-  } catch (error) {
-    throw Error(error);
-  }
-};
-module.exports = { createThread, getAllForumThreads };
+module.exports = { createThread };
